6-size.c

+++ 0x00-hello_world/6-size.c
@@ -1,8 +1,11 @@
 #include <stdio.h>
+
 /**
- * main - A program that prints the size of various compture types
- * return 0 (success)
+ * main - Entry point
+ *
+ * Return: Always 0 (success)
  */
+
 int main(void)
 {
 	char a;
@@ -11,10 +14,11 @@ int main(void)
 	long long int d;
 	float f;
 
-printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(a));
-printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(b));
-printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(c));
-printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
-printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
-return (0);
+	printf("Size of a char: %lu byte(s)\n", (sizeof(a));
+	printf("Size of an int: %lu byte(s)\n", (sizeof(b));
+	printf("Size of a long int: %lu byte(s)\n", (sizeof(c));
+	printf("Size of a long long int: %lu byte(s)\n", (sizeof(d));
+	printf("Size of a float: %lu byte(s)\n", (sizeof(f));
+	return (0);
+
 }
